requirements.txt
# Web Framework
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
requests>=2.31.0

# Machine Learning
torch>=2.2.0
transformers>=4.38.0
accelerate>=0.27.0

# PowerPoint Processing
python-pptx>=0.6.23
pandas>=2.2.0

# Utilities
python-dotenv>=1.0.1
tqdm>=4.66.0
numpy>=1.26.0 